refactor(finance): migrate domain to @modules/finance, kill repos
- routers + middlewares in modules/finance/
- 5 finance repositories deleted (transactions, bank-accounts,
  credit-cards, credit-card-statements, financial-settings)
- routers query context.db direct via neverthrow + drizzle
- ownership/refs via composable oRPC middlewares
- transactions split: crud, list, status, bulk, suggestions
- financial-settings stays in @modules/account (account-scoped config)
