+++ apps/web/package.json
@@ -42,9 +42,11 @@
       "@maskito/core": "catalog:ui",
       "@maskito/kit": "catalog:ui",
       "@maskito/react": "catalog:ui",
+      "@modules/account": "workspace:*",
       "@modules/agents": "workspace:*",
       "@modules/billing": "workspace:*",
       "@modules/classification": "workspace:*",
+      "@modules/finance": "workspace:*",
       "@orpc/client": "catalog:orpc",
       "@orpc/json-schema": "catalog:orpc",
       "@orpc/openapi": "catalog:orpc",
