--- apps/web/src/integrations/orpc/router/transactions.ts
@@ -1,745 +1,0 @@
-import { and, eq, inArray } from "drizzle-orm";
-import {
-   ConditionGroup,
-   evaluateConditionGroup,
-} from "@f-o-t/condition-evaluator";
-import {
-   bulkCreateTransactions,
-   bulkMarkTransactionsAsPaid,
-   cancelTransaction,
-   createTransaction,
-   createTransactionItems,
-   deleteTransaction,
-   ensureTransactionOwnership,
-   getTransactionsSummary,
-   getTransactionWithTag,
-   listTransactions,
-   markTransactionAsPaid,
-   markTransactionAsUnpaid,
-   reactivateTransaction,
-   replaceTransactionItems,
-   updateTransaction,
-   updateTransactionCategory,
-   updateTransactionTag,
-   validateTransactionReferences,
-} from "@core/database/repositories/transactions-repository";
-import { ensureBankAccountOwnership } from "@core/database/repositories/bank-accounts-repository";
-import { enforceCostCenterPolicy } from "@core/database/repositories/financial-settings-repository";
-import {
-   createTransactionSchema,
-   transactions,
-   updateTransactionSchema,
-} from "@core/database/schemas/transactions";
-import { createEmitFn } from "@packages/events/emit";
-import { emitFinanceStatementImported } from "@packages/events/finance";
-import { WebAppError } from "@core/logging/errors";
-import { z } from "zod";
-import { enqueueClassifyTransactionsBatchWorkflow } from "@modules/classification/workflows/classification-workflow";
-import { withCreditEnforcement } from "../middlewares/credit-enforcement";
-import { protectedProcedure } from "../server";
-
-const idSchema = z.object({ id: z.string().uuid() });
-
-const tagAndItemsSchema = z.object({
-   tagId: z.string().uuid().nullable().optional(),
-   items: z
-      .array(
-         z.object({
-            serviceId: z.string().uuid().nullable().optional(),
-            description: z.string().max(500).nullable().optional(),
-            quantity: z.string(),
-            unitPrice: z.string(),
-         }),
-      )
-      .optional()
-      .default([]),
-});
-
-const filterSchema = z
-   .object({
-      type: z.enum(["income", "expense", "transfer"]).optional(),
-      bankAccountId: z.string().uuid().optional(),
-      categoryId: z.string().uuid().optional(),
-      tagId: z.string().uuid().optional(),
-      contactId: z.string().uuid().optional(),
-      dateFrom: z
-         .string()
-         .regex(/^\d{4}-\d{2}-\d{2}$/)
-         .optional(),
-      dateTo: z
-         .string()
-         .regex(/^\d{4}-\d{2}-\d{2}$/)
-         .optional(),
-      search: z.string().max(100).optional(),
-      creditCardId: z.string().uuid().optional(),
-      uncategorized: z.boolean().optional(),
-      paymentMethod: z.string().optional(),
-      status: z
-         .union([
-            z.enum(["pending", "paid", "cancelled"]),
-            z.array(z.enum(["pending", "paid", "cancelled"])),
-         ])
-         .optional(),
-      dueDateFrom: z
-         .string()
-         .regex(/^\d{4}-\d{2}-\d{2}$/)
-         .optional(),
-      dueDateTo: z
-         .string()
-         .regex(/^\d{4}-\d{2}-\d{2}$/)
-         .optional(),
-      overdueOnly: z.boolean().optional(),
-      view: z
-         .enum(["all", "payable", "receivable", "settled", "cancelled"])
-         .optional(),
-      page: z.number().int().positive().default(1),
-      pageSize: z.number().int().positive().max(100).default(20),
-      conditionGroup: ConditionGroup.optional(),
-   })
-   .optional();
-
-export const create = protectedProcedure
-   .input(
-      createTransactionSchema
-         .merge(tagAndItemsSchema)
-         .merge(z.object({ autoCategorize: z.boolean().default(false) })),
-   )
-   .handler(async ({ context, input }) => {
-      const { tagId, items, autoCategorize, ...data } = input;
-      await enforceCostCenterPolicy(context.db, context.teamId, tagId);
-      await validateTransactionReferences(context.db, context.teamId, {
-         bankAccountId: data.bankAccountId,
-         destinationBankAccountId: data.destinationBankAccountId,
-         categoryId: data.type === "transfer" ? null : data.categoryId,
-         tagId,
-         contactId: data.contactId,
-         date: data.date,
-      });
-
-      const txData =
-         data.type === "transfer"
-            ? { ...data, categoryId: null, type: "transfer" as const }
-            : data;
-
-      const transaction = (
-         await createTransaction(
-            context.db,
-            context.teamId,
-            txData,
-            tagId ?? undefined,
-         )
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-
-      if (items.length > 0 && transaction) {
-         (
-            await createTransactionItems(
-               context.db,
-               transaction.id,
-               context.teamId,
-               items,
-            )
-         ).match(
-            (v) => v,
-            (e) => {
-               throw e;
-            },
-         );
-      }
-
-      if (
-         autoCategorize &&
-         transaction &&
-         !input.categoryId &&
-         (input.type === "income" || input.type === "expense")
-      ) {
-         await enqueueClassifyTransactionsBatchWorkflow(
-            context.workflowClient,
-            {
-               teamId: context.teamId,
-               transactionIds: [transaction.id],
-            },
-         );
-      }
-
-      return transaction;
-   });
-
-export const getAll = protectedProcedure
-   .input(filterSchema)
-   .handler(async ({ context, input }) => {
-      return (
-         await listTransactions(context.db, {
-            teamId: context.teamId,
-            ...input,
-         })
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
-
-export const getSummary = protectedProcedure
-   .input(filterSchema)
-   .handler(async ({ context, input }) => {
-      return (
-         await getTransactionsSummary(context.db, {
-            teamId: context.teamId,
-            ...input,
-         })
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
-
-export const getById = protectedProcedure
-   .input(idSchema)
-   .handler(async ({ context, input }) => {
-      await ensureTransactionOwnership(context.db, input.id, context.teamId);
-      return (await getTransactionWithTag(context.db, input.id)).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
-
-export const update = protectedProcedure
-   .input(
-      idSchema
-         .merge(updateTransactionSchema)
-         .merge(tagAndItemsSchema.partial()),
-   )
-   .handler(async ({ context, input }) => {
-      const existing = await ensureTransactionOwnership(
-         context.db,
-         input.id,
-         context.teamId,
-      );
-      if ("tagId" in input) {
-         await enforceCostCenterPolicy(context.db, context.teamId, input.tagId);
-      }
-      if (
-         input.bankAccountId ||
-         input.destinationBankAccountId ||
-         input.categoryId ||
-         input.tagId ||
-         input.contactId
-      ) {
-         await validateTransactionReferences(context.db, context.teamId, {
-            bankAccountId: input.bankAccountId ?? existing.bankAccountId,
-            destinationBankAccountId: input.destinationBankAccountId,
-            categoryId: input.categoryId,
-            tagId: input.tagId,
-            contactId: input.contactId,
-            date: input.date ?? existing.date,
-         });
-      }
-      const { id, tagId, items, ...data } = input;
-      const result = (
-         await updateTransaction(context.db, id, data, tagId)
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-      if (items !== undefined) {
-         (
-            await replaceTransactionItems(context.db, id, context.teamId, items)
-         ).match(
-            (v) => v,
-            (e) => {
-               throw e;
-            },
-         );
-      }
-      return result;
-   });
-
-export const remove = protectedProcedure
-   .input(idSchema)
-   .handler(async ({ context, input }) => {
-      await ensureTransactionOwnership(context.db, input.id, context.teamId);
-      (await deleteTransaction(context.db, input.id)).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-      return { success: true };
-   });
-
-export const importStatement = withCreditEnforcement(
-   "finance.statement_imported",
-)
-   .input(
-      z.object({
-         bankAccountId: z.string().uuid(),
-         format: z.enum(["csv", "xlsx", "ofx"]),
-         transactions: z
-            .array(
-               z.object({
-                  name: z.string().max(500).optional(),
-                  type: z.enum(["income", "expense"]),
-                  amount: z.string().regex(/^-?\d+(\.\d+)?$/),
-                  date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/),
-                  description: z.string().max(1000).optional(),
-                  categoryId: z.string().uuid().optional(),
-                  paymentMethod: z
-                     .enum([
-                        "pix",
-                        "credit_card",
-                        "debit_card",
-                        "boleto",
-                        "cash",
-                        "transfer",
-                        "other",
-                        "cheque",
-                        "automatic_debit",
-                     ])
-                     .optional(),
-               }),
-            )
-            .min(1)
-            .max(1000),
-         autoCategorize: z.boolean().default(false),
-      }),
-   )
-   .handler(async ({ context, input }) => {
-      await ensureBankAccountOwnership(
-         context.db,
-         input.bankAccountId,
-         context.teamId,
-      );
-
-      const uniqueCategoryIds = [
-         ...new Set(
-            input.transactions
-               .map((t) => t.categoryId)
-               .filter((id): id is string => !!id),
-         ),
-      ];
-      for (const categoryId of uniqueCategoryIds) {
-         await validateTransactionReferences(context.db, context.teamId, {
-            categoryId,
-         });
-      }
-
-      const rows = input.transactions.map((t) => ({
-         bankAccountId: input.bankAccountId,
-         name: t.name ?? null,
-         type: t.type,
-         amount: t.amount,
-         date: t.date,
-         description: t.description ?? null,
-         categoryId: t.categoryId ?? null,
-         paymentMethod: t.paymentMethod ?? null,
-      }));
-
-      const inserted = (
-         await bulkCreateTransactions(context.db, context.teamId, rows)
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-
-      if (input.autoCategorize) {
-         const idsToClassify = inserted
-            .filter(
-               (tx) =>
-                  !tx.categoryId &&
-                  (tx.type === "income" || tx.type === "expense"),
-            )
-            .map((tx) => tx.id);
-         if (idsToClassify.length > 0) {
-            await enqueueClassifyTransactionsBatchWorkflow(
-               context.workflowClient,
-               {
-                  teamId: context.teamId,
-                  transactionIds: idsToClassify,
-               },
-            );
-         }
-      }
-
-      try {
-         const emit = createEmitFn(context.db, context.posthog);
-         await emitFinanceStatementImported(
-            emit,
-            {
-               organizationId: context.organizationId,
-               userId: context.userId,
-               teamId: context.teamId,
-            },
-            {
-               bankAccountId: input.bankAccountId,
-               format: input.format,
-               rowCount: rows.length,
-            },
-         );
-      } catch {}
-
-      return { imported: rows.length };
-   });
-
-function normalizeAmount(amount: string): string {
-   return Number.parseFloat(amount).toFixed(2);
-}
-
-export const checkDuplicates = protectedProcedure
-   .input(
-      z.object({
-         bankAccountId: z.string().uuid(),
-         transactions: z
-            .array(
-               z.object({
-                  date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/),
-                  amount: z.string().regex(/^-?\d+(\.\d+)?$/),
-                  type: z.enum(["income", "expense"]),
-               }),
-            )
-            .min(1)
-            .max(1000),
-      }),
-   )
-   .handler(async ({ context, input }) => {
-      await ensureBankAccountOwnership(
-         context.db,
-         input.bankAccountId,
-         context.teamId,
-      );
-
-      const existing = await context.db
-         .select({
-            date: transactions.date,
-            amount: transactions.amount,
-            type: transactions.type,
-         })
-         .from(transactions)
-         .where(
-            and(
-               eq(transactions.bankAccountId, input.bankAccountId),
-               eq(transactions.teamId, context.teamId),
-               inArray(
-                  transactions.date,
-                  input.transactions.map((t) => t.date),
-               ),
-            ),
-         );
-
-      return input.transactions.map((t) => {
-         const normalizedAmt = normalizeAmount(t.amount);
-         return existing.some((ex) => {
-            const result = evaluateConditionGroup(
-               {
-                  id: "dup",
-                  operator: "AND",
-                  scoringMode: "weighted",
-                  threshold: 0.75,
-                  conditions: [
-                     {
-                        id: "date-group",
-                        operator: "AND",
-                        weight: 3,
-                        conditions: [
-                           {
-                              id: "date",
-                              type: "string",
-                              field: "date",
-                              operator: "eq",
-                              value: ex.date,
-                           },
-                        ],
-                     },
-                     {
-                        id: "amount-group",
-                        operator: "AND",
-                        weight: 3,
-                        conditions: [
-                           {
-                              id: "amount",
-                              type: "string",
-                              field: "amount",
-                              operator: "eq",
-                              value: normalizeAmount(ex.amount),
-                           },
-                        ],
-                     },
-                     {
-                        id: "type-group",
-                        operator: "AND",
-                        weight: 2,
-                        conditions: [
-                           {
-                              id: "type",
-                              type: "string",
-                              field: "type",
-                              operator: "eq",
-                              value: ex.type,
-                           },
-                        ],
-                     },
-                  ],
-               },
-               { data: { date: t.date, amount: normalizedAmt, type: t.type } },
-            );
-            return result.passed;
-         });
-      });
-   });
-
-export const importBulk = protectedProcedure
-   .input(
-      z.object({
-         transactions: z
-            .array(createTransactionSchema.merge(tagAndItemsSchema))
-            .min(1)
-            .max(500),
-         autoCategorize: z.boolean().default(false),
-      }),
-   )
-   .handler(async ({ context, input }) => {
-      let imported = 0;
-      const idsToClassify: string[] = [];
-      for (const t of input.transactions) {
-         const { tagId, items: _items, ...data } = t;
-         await validateTransactionReferences(context.db, context.teamId, {
-            bankAccountId: data.bankAccountId,
-            destinationBankAccountId: data.destinationBankAccountId,
-            categoryId: data.categoryId,
-            tagId,
-            date: data.date,
-         });
-         const transaction = (
-            await createTransaction(
-               context.db,
-               context.teamId,
-               data,
-               tagId ?? undefined,
-            )
-         ).match(
-            (v) => v,
-            (e) => {
-               throw e;
-            },
-         );
-         if (
-            input.autoCategorize &&
-            transaction &&
-            !data.categoryId &&
-            (data.type === "income" || data.type === "expense")
-         ) {
-            idsToClassify.push(transaction.id);
-         }
-         imported++;
-      }
-      if (idsToClassify.length > 0) {
-         await enqueueClassifyTransactionsBatchWorkflow(
-            context.workflowClient,
-            {
-               teamId: context.teamId,
-               transactionIds: idsToClassify,
-            },
-         );
-      }
-      return { imported, skipped: 0 };
-   });
-
-export const acceptSuggestedCategory = protectedProcedure
-   .input(z.object({ id: z.string().uuid() }))
-   .handler(async ({ context, input }) => {
-      const tx = await ensureTransactionOwnership(
-         context.db,
-         input.id,
-         context.teamId,
-      );
-      if (!tx.suggestedCategoryId) {
-         throw WebAppError.badRequest(
-            "Nenhuma sugestão de categoria disponível.",
-         );
-      }
-      (
-         await updateTransactionCategory(context.db, input.id, {
-            categoryId: tx.suggestedCategoryId,
-            suggestedCategoryId: null,
-         })
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-      return { ok: true };
-   });
-
-export const dismissSuggestedCategory = protectedProcedure
-   .input(z.object({ id: z.string().uuid() }))
-   .handler(async ({ context, input }) => {
-      await ensureTransactionOwnership(context.db, input.id, context.teamId);
-      (
-         await updateTransactionCategory(context.db, input.id, {
-            suggestedCategoryId: null,
-         })
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-      return { ok: true };
-   });
-
-export const acceptSuggestedTag = protectedProcedure
-   .input(z.object({ id: z.string().uuid() }))
-   .handler(async ({ context, input }) => {
-      const tx = await ensureTransactionOwnership(
-         context.db,
-         input.id,
-         context.teamId,
-      );
-      if (!tx.suggestedTagId) {
-         throw WebAppError.badRequest(
-            "Nenhuma sugestão de centro de custo disponível.",
-         );
-      }
-      (
-         await updateTransactionTag(context.db, input.id, {
-            tagId: tx.suggestedTagId,
-            suggestedTagId: null,
-         })
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-      return { ok: true };
-   });
-
-export const dismissSuggestedTag = protectedProcedure
-   .input(z.object({ id: z.string().uuid() }))
-   .handler(async ({ context, input }) => {
-      await ensureTransactionOwnership(context.db, input.id, context.teamId);
-      (
-         await updateTransactionTag(context.db, input.id, {
-            suggestedTagId: null,
-         })
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-      return { ok: true };
-   });
-
-export const markAsPaid = protectedProcedure
-   .input(
-      z.object({
-         id: z.string().uuid(),
-         paidDate: z
-            .string()
-            .regex(/^\d{4}-\d{2}-\d{2}$/)
-            .optional(),
-         bankAccountId: z.string().uuid().nullable().optional(),
-      }),
-   )
-   .handler(async ({ context, input }) => {
-      return (
-         await markTransactionAsPaid(context.db, input.id, context.teamId, {
-            paidDate: input.paidDate,
-            bankAccountId: input.bankAccountId ?? undefined,
-         })
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
-
-export const markAsUnpaid = protectedProcedure
-   .input(idSchema)
-   .handler(async ({ context, input }) => {
-      return (
-         await markTransactionAsUnpaid(context.db, input.id, context.teamId)
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
-
-export const cancel = protectedProcedure
-   .input(idSchema)
-   .handler(async ({ context, input }) => {
-      return (
-         await cancelTransaction(context.db, input.id, context.teamId)
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
-
-export const reactivate = protectedProcedure
-   .input(z.object({ id: z.string().uuid(), paid: z.boolean().default(false) }))
-   .handler(async ({ context, input }) => {
-      return (
-         await reactivateTransaction(
-            context.db,
-            input.id,
-            context.teamId,
-            input.paid,
-         )
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
-
-export const bulkMarkAsPaid = protectedProcedure
-   .input(
-      z.object({
-         ids: z.array(z.string().uuid()).min(1).max(500),
-         paidDate: z
-            .string()
-            .regex(/^\d{4}-\d{2}-\d{2}$/)
-            .optional(),
-         bankAccountId: z.string().uuid().nullable().optional(),
-      }),
-   )
-   .handler(async ({ context, input }) => {
-      return (
-         await bulkMarkTransactionsAsPaid(
-            context.db,
-            input.ids,
-            context.teamId,
-            {
-               paidDate: input.paidDate,
-               bankAccountId: input.bankAccountId ?? undefined,
-            },
-         )
-      ).match(
-         (v) => v,
-         (e) => {
-            throw e;
-         },
-      );
-   });
