+++ modules/finance/src/router/transactions-index.ts
@@ -1,0 +1,25 @@
+export {
+   create,
+   getById,
+   update,
+   remove,
+} from "@modules/finance/router/transactions";
+export { getAll, getSummary } from "@modules/finance/router/transactions-list";
+export {
+   acceptSuggestedCategory,
+   acceptSuggestedTag,
+   dismissSuggestedCategory,
+   dismissSuggestedTag,
+} from "@modules/finance/router/transactions-suggestions";
+export {
+   bulkMarkAsPaid,
+   cancel,
+   markAsPaid,
+   markAsUnpaid,
+   reactivate,
+} from "@modules/finance/router/transactions-status";
+export {
+   checkDuplicates,
+   importBulk,
+   importStatement,
+} from "@modules/finance/router/transactions-bulk";
