+++ modules/finance/tsconfig.json
@@ -1,0 +1,21 @@
+{
+   "extends": "@tooling/typescript/core.json",
+   "compilerOptions": {
+      "composite": false,
+      "declaration": false,
+      "declarationMap": false,
+      "noEmit": true,
+      "jsx": "react-jsx",
+      "paths": {
+         "@core/database": ["../../core/database/src/index.ts"],
+         "@core/database/*": ["../../core/database/src/*"],
+         "@core/logging": ["../../core/logging/src/logger.ts"],
+         "@core/logging/*": ["../../core/logging/src/*"],
+         "@core/orpc/*": ["../../core/orpc/src/*"],
+         "@core/redis/*": ["../../core/redis/src/*"],
+         "@modules/classification/*": ["../classification/src/*"],
+         "@packages/events/*": ["../../packages/events/src/*"]
+      }
+   },
+   "include": ["src"]
+}
